rename: errorUtil.js 파일명 오타 수정 (#137)
