+++ src/pages/login/index.js
@@ -1,7 +1,7 @@
 import { loginUser } from '../../apis/userApi';
 import { route, url } from '../../router';
 import { authManager } from '../../services/auth';
-import { errorHendler } from '../../utils/errorUtile';
+import { errorHendler } from '../../utils/errorUtil';
 import { Form } from '../../components/common/form';
 import { FORM_FIELDS, FORM_BUTTONS } from './formFieldDatas';
