+++ src/pages/signup/index.js
@@ -1,9 +1,8 @@
 import { postUserData } from '../../apis/userApi';
 import { route, url } from '../../router';
-import { errorHendler } from '../../utils/errorUtile';
+import { errorHendler } from '../../utils/errorUtil';
 import { Form } from '../../components/common/form';
 import { FORM_FIELDS, FORM_BUTTONS } from './formFieldDatas';
-import AvatarImg from '/public/avatar.svg';
 
 export default class SignUpPage {
 	#contentsElement;
